+++ lib/ex/parser/subscanners.py
@@ -35,6 +35,7 @@ from .scanner_command_set import scan_command_set
 from .scanner_command_set_local import scan_command_set_local
 from .scanner_command_shell import scan_command_shell
 from .scanner_command_shell_out import scan_command_shell_out
+from .scanner_command_smap import scan_command_smap
 from .scanner_command_substitute import scan_command_substitute
 from .scanner_command_tab_first import scan_command_tab_first
 from .scanner_command_tab_last import scan_command_tab_last
@@ -94,6 +95,7 @@ patterns[r's(?:ubstitute)?(?=[%&:/=]|$)'] = scan_command_substitute
 patterns[r'se(?:t)?(?=$|\s)'] = scan_command_set
 patterns[r'setl(?:ocal)?'] = scan_command_set_local
 patterns[r'sh(?:ell)?'] = scan_command_shell
+patterns[r'sm(?:ap)?'] = scan_command_smap
 patterns[r'tabfir(?:st)?'] = scan_command_tab_first
 patterns[r'tabl(?:ast)?'] = scan_command_tab_last
 patterns[r'tabn(?:ext)?'] = scan_command_tab_next